fix: scrollLeft calculation logic
