+++ web/components/gantt-chart/helpers/index.ts
@@ -1,3 +1,2 @@
 export * from "./add-block";
-export * from "./block-structure";
 export * from "./draggable";
