+++ web/components/gantt-chart/index.ts
@@ -3,5 +3,5 @@ export * from "./chart";
 export * from "./helpers";
 export * from "./hooks";
 export * from "./root";
-export * from "./types";
 export * from "./sidebar";
+export * from "./types";
