+++ web/components/issues/issue-layouts/gantt/base-gantt-root.tsx
@@ -5,12 +5,9 @@ import { observer } from "mobx-react-lite";
 import { useIssues, useUser } from "hooks/store";
 // components
 import { GanttQuickAddIssueForm, IssueGanttBlock } from "components/issues";
-import {
-  GanttChartRoot,
-  IBlockUpdateData,
-  renderIssueBlocksStructure,
-  IssueGanttSidebar,
-} from "components/gantt-chart";
+import { GanttChartRoot, IBlockUpdateData, IssueGanttSidebar } from "components/gantt-chart";
+// helpers
+import { renderIssueBlocksStructure } from "helpers/issue.helper";
 // types
 import { TIssue, TUnGroupedIssues } from "@plane/types";
 import { ICycleIssues, ICycleIssuesFilter } from "store/issue/cycle";
