+++ web/helpers/issue.helper.ts
@@ -3,6 +3,7 @@ import { v4 as uuidv4 } from "uuid";
 import { orderArrayBy } from "helpers/array.helper";
 // types
 import { TIssue, TIssueGroupByOptions, TIssueLayouts, TIssueOrderByOptions, TIssueParams } from "@plane/types";
+import { IGanttBlock } from "components/gantt-chart";
 // constants
 import { ISSUE_DISPLAY_FILTERS_BY_LAYOUT } from "constants/issue";
 
@@ -132,3 +133,12 @@ export const createIssuePayload: (projectId: string, formData: Partial<TIssue>) 
 
   return payload;
 };
+
+export const renderIssueBlocksStructure = (blocks: TIssue[]): IGanttBlock[] =>
+  blocks?.map((block) => ({
+    data: block,
+    id: block.id,
+    sort_order: block.sort_order,
+    start_date: block.start_date ? new Date(block.start_date) : null,
+    target_date: block.target_date ? new Date(block.target_date) : null,
+  }));
